feat: add new character sheet page
Refactoring implemented during the addition of the new CharacterSheetPage, considering the product
Smart X Sheet for all the sheet's functionality.

During the finalization of this page, while I was leaving the necessary points for CreateCharacterSheet,
the need to treat the Smart X Sheet with greater importance within the application was realized.
Several decisions were considered to modularize, isolate, design, and architect what it would look like.
Then the new refactoring was implemented just before the new CharacterSheetPage was finalized.